Refine return types for all methods
Previously, all methods returned a generic `Response`. However, only
certain variants are possible responses to each command. This forces
users to manually match on the returned types, even when that shouldn't
be necessary. This patch does the matching _inside_ the library, so that
only the expected return value is exposed. If an incorrect variant
(according to the protocol) is returned, an error is returned instead.